[PORT FROM MAIN] wifi: wlan_prov for BCM
BZ: 55429

This is a DO NOT UPSTREAM patch.

This patch enables wlan_prov for Broadcom solution
wlan_prov read the mac address from Chaabi
and it creates (or updates) the mac file according to it.
If Chaabi support is not available, a random mac address is used.

Feature   : Wlan
Category  : Differentiation
Purpose   : AOSP-ext
Origin    : Internal
Upstream-candidate: no

Change-Id: I0f56c6b8f93370ce6d995c4982c876964093e4ab
Orig-Change-Id: If652d8de0199a75953ec20d5cbeed05229332126
Signed-off-by: Marco Sinigaglia <marcox.sinigaglia@intel.com>