use TARGET_COPY_OUT_VENDOR
The AOSP provides the TARGET_COPY_OUT_VENDOR variable to allow the
vendor specific files to be easily moved to a dedicated vendor
partition.  By default, TARGET_COPY_OUT_VENDOR is set to
"system/vendor".  If a dedicated vendor partition is in used,
TARGET_COPY_OUT_VENDOR is set to "vendor'.

This patch replaces the hard-coded "system/vendor" with
$(TARGET_COPY_OUT_VENDOR).

Change-Id: Iebc81622aa8ecdd0d14c45118abdbacdb7467995
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com/339558
Reviewed-by: Fradin, Benoit <benoit.fradin@intel.com>
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>